refactor manage related pages
