+++ src/stories/containers/users/create-account/create-account.tsx
@@ -5,7 +5,7 @@ import CloseButton from '../../../components/close-button/close-button';
 import { useThemeContext } from '../../../../core/context/ThemeContext';
 import { CustomButton } from '../../../components/custom-button/custom-button';
 import TextInput from '../../../components/text-input/text-input';
-import { ButtonWrapper, Container, Wrapper, Form } from '../../auth/login/login';
+import { ButtonWrapper, Wrapper, Form } from '../../auth/login/login';
 import { useCreateAccountMvvm } from './create-account.mvvm';
 import { useRouter } from 'next/router';
 import { useAuthContext } from '../../../../core/context/AuthContext';
@@ -108,6 +108,25 @@ export default () => {
     </Wrapper>
   );
 };
+
+export const Container = styled.div<{ isLight?: boolean }>(({ isLight }) => ({
+  display: 'flex',
+  flexDirection: 'column',
+  alignItems: 'center',
+  padding: 24,
+  width: 343,
+  background: isLight ? '#FFFFFF' : '#10191F',
+  boxShadow: isLight
+    ? '0px 20px 40px rgba(219, 227, 237, 0.4), 0px 1px 3px rgba(190, 190, 190, 0.25)'
+    : '0px 20px 40px -40px rgba(7, 22, 40, 0.4), 0px 1px 3px rgba(30, 23, 23, 0.25)',
+
+  borderRadius: '6px',
+  '@media (min-width: 834px)': {
+    padding: '40px 64px',
+    width: '484px',
+    top: 128,
+  },
+}));
 
 const Title = styled.h1<{ isLight: boolean }>(({ isLight }) => ({
   margin: '24px 0 16px 0',
