+++ src/stories/containers/users/users-manager/manager-tabs.enum.ts
@@ -1,0 +1,4 @@
+export enum ManagerTabs {
+  PROFILE = 0,
+  MANAGER = 1,
+}
